+++ test/test_helper.exs
@@ -1,2 +1,3 @@
 {:ok, _} = Application.ensure_all_started(:ecto_sql)
+File.mkdir_p!("tmp")
 ExUnit.start(exclude: [:s3, :ecto])